add notification interval chooser

+++ webapp/lib/f_tracking/services/tracker_provider.dart
@@ -1,3 +1,4 @@
+import 'package:cosmos_notifier/api/protobuf/dart/google/protobuf/duration.pb.dart' as pb;
 import 'package:cosmos_notifier/api/protobuf/dart/google/protobuf/empty.pb.dart';
 import 'package:cosmos_notifier/api/protobuf/dart/tracker_service.pb.dart';
 import 'package:cosmos_notifier/config.dart';
@@ -6,7 +7,6 @@ import 'package:cosmos_notifier/f_tracking/services/state/tracker_row.dart';
 import 'package:cosmos_notifier/f_tracking/services/tracker_service.dart';
 import 'package:fixnum/fixnum.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:cosmos_notifier/api/protobuf/dart/google/protobuf/duration.pb.dart' as pb;
 
 final trackerFutureProvider = FutureProvider((ref) async {
   await ref.read(trackerNotifierProvider.notifier).getTrackers();
@@ -16,11 +16,32 @@ final trackerNotifierProvider = StateNotifierProvider<TrackerNotifier, List<Trac
   return TrackerNotifier(trackerService, ref.read(messageProvider.notifier));
 });
 
+final showAddTrackerButtonProvider = Provider((ref) {
+  return ref.watch(trackerNotifierProvider.notifier).hasUnsavedChanges;
+});
+
 class TrackerNotifier extends StateNotifier<List<TrackerRow>> {
   TrackerService trackerService;
   MessageNotifier messageNotifier;
 
   TrackerNotifier(this.trackerService, this.messageNotifier) : super([]);
+
+  pb.Duration _getDefaultNotificationInterval() {
+    var lastModifiedTracker = state.where((trackerRow) => trackerRow.updatedAt != null).reduce((value, element) {
+      if (value.updatedAt != null && element.updatedAt != null && value.updatedAt!.isBefore(element.updatedAt!)) {
+        return value;
+      }
+      return element;
+    });
+    if (lastModifiedTracker.updatedAt == null) {
+      return pb.Duration(seconds: Int64(60 * 60 * 24)); // 1 day by default
+    }
+    return lastModifiedTracker.notificationInterval;
+  }
+
+  bool get hasUnsavedChanges {
+    return state.where((trackerRow) => !trackerRow.isSaved).isNotEmpty;
+  }
 
   Future<void> getTrackers() async {
     var response = await trackerService.getTrackers(Empty());
@@ -32,40 +53,35 @@ class TrackerNotifier extends StateNotifier<List<TrackerRow>> {
           address: tracker.address,
           notificationInterval: tracker.notificationInterval,
           chatId: tracker.discordChannelId,
-          isSaved: true,
+          updatedAt: tracker.updatedAt.toDateTime(),
         )
       ];
     }
-    state = [...state, TrackerRow(id: Int64(0), address: "", notificationInterval: pb.Duration(), chatId: Int64(0), isSaved: false, isAddRow: true)];
   }
 
   void addTracker() {
-    if (state.last.isAddRow) {
+    if (hasUnsavedChanges) {
       return;
     }
-    state = [...state, TrackerRow(id: Int64(0), address: "", notificationInterval: pb.Duration(), chatId: Int64(0), isSaved: false)];
+    state = [
+      ...state,
+      TrackerRow(id: Int64(0), address: "", notificationInterval: _getDefaultNotificationInterval(), chatId: Int64(0), updatedAt: null),
+    ];
   }
 
   Future<void> updateTracker(TrackerRow tracker) async {
-    if (tracker.isAddRow) {
+    if (!tracker.isSaved) {
       if (tracker.address.isNotEmpty) {
         try {
           var response = await trackerService.isAddressValid(IsAddressValidRequest(address: tracker.address));
           tracker = tracker.copyWith(isAddressValid: response.isValid);
-          state = [
-            for (final oldTrackerRow in state)
-              if (oldTrackerRow.id == tracker.id)
-                tracker
-              else
-                oldTrackerRow,
-          ];
         } catch (e) {
           messageNotifier.sendMsg(error: e.toString());
           return;
         }
       }
       // if (tracker.address.isNotEmpty && !tracker.notificationInterval.seconds.isZero && tracker.chatId.toInt() > 0) {
-      if (tracker.isAddressValid && !tracker.notificationInterval.seconds.isZero) {
+      if (tracker.address.isNotEmpty && tracker.isAddressValid && !tracker.notificationInterval.seconds.isZero) {
         try {
           var response = await trackerService.addTracker(AddTrackerRequest(
             address: tracker.address,
@@ -75,12 +91,38 @@ class TrackerNotifier extends StateNotifier<List<TrackerRow>> {
           // TODO: fix state
         } catch (e) {
           messageNotifier.sendMsg(error: e.toString());
+          return;
         }
+      }
+      state = _updateTrackerRow(tracker);
+    } else {
+      try {
+        await trackerService.updateTracker(UpdateTrackerRequest(
+          trackerId: tracker.id,
+          notificationInterval: tracker.notificationInterval,
+          discordChannelId: tracker.chatId,
+          // TODO: telegram chat id
+
+        ));
+        state = _updateTrackerRow(tracker);
+      } catch (e) {
+        messageNotifier.sendMsg(error: e.toString());
       }
     }
   }
 
+  List<TrackerRow> _updateTrackerRow(TrackerRow tracker) {
+    return [
+      for (final oldTrackerRow in state)
+        if (oldTrackerRow.id == tracker.id) tracker else oldTrackerRow,
+    ];
+  }
+
   Future<void> deleteTracker(TrackerRow tracker) async {
+    if (!tracker.isSaved) {
+      state = state.where((element) => element.id != tracker.id).toList();
+      return;
+    }
     try {
       await trackerService.deleteTracker(DeleteTrackerRequest(trackerId: tracker.id));
       state = state.where((element) => element.id != tracker.id).toList();
